--- 1Week_Mission/src/main/java/com/example/mission/app/post/entity/PostKeyword.java
@@ -1,21 +1,0 @@
-package com.example.mission.app.post.entity;
-
-import javax.persistence.Entity;
-
-import com.example.mission.app.base.entity.BaseEntity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@SuperBuilder
-@ToString(callSuper = true)
-public class PostKeyword extends BaseEntity {
-	private String content;
-}
